+++ src/main/java/org/opensearch/neuralsearch/query/HybridSubQueryScorer.java
@@ -5,6 +5,7 @@
 package org.opensearch.neuralsearch.query;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import org.apache.lucene.search.Scorable;
 
 import java.io.IOException;
@@ -13,6 +14,7 @@ import java.util.Arrays;
 /**
  * Scorer implementation for Hybrid Query. This object is light and expected to be re-used between different doc ids
  */
+@EqualsAndHashCode(callSuper = true)
 @Data
 public class HybridSubQueryScorer extends Scorable {
     // array of scores from all sub-queries for a single doc id